[sapphire] Issue #8: feat: Move payout methods to new service
